Conflicts:
	groupwareWebApp/target/m2e-wtp/web-resources/META-INF/maven/com.bit/groupware/pom.properties

+++ groupwareWebApp/src/main/java/com/bit/groupware/controller/approval/ApprovalAjaxController.java
@@ -52,15 +52,17 @@ public class ApprovalAjaxController {
 	public int approvalAjax(ApprovalVO approval, 
 			@RequestParam(value="tmpNo",required=false,defaultValue="0") int tmpNo, 
 			@RequestParam int receiverNo, 
-			@RequestParam(value="deleteAppr", required=false, defaultValue="0") int deleteAppr , 
+			@RequestParam(value="deleteAppr", required=false, defaultValue="0") int deleteAppr ,
+			@RequestParam(value="reApprDelete", required=false, defaultValue="0") int reApprDelete,
 			HttpSession session, 
 			Principal principal) throws Exception {
 		//approval => validDate, urgency, apprTitle, apprContent,  apprFinalStatus
 		
 		List<Integer> apprNos =new ArrayList<Integer>();
 		apprNos.add(deleteAppr); 
-		approvalService.removeApproval(apprNos);  
-		
+		if(reApprDelete==1) {
+			approvalService.removeApproval(apprNos);  
+		}
 		EmployeeVO employee = new EmployeeVO();
 		employee.setEmpNo(principal.getName());
 		
@@ -76,7 +78,6 @@ public class ApprovalAjaxController {
 				approval.addApprovalFile(approvalFile);
 			}
 		}
-		System.out.println(session.getServletContext().getRealPath("/") +"zzzzzzzzzzzzzzzzz");
 		approvalService.registerApproval(approval, receiverNo);
 		
 		return approval.getApprFinalStatus();
